Manca 3za pagina

+++ m2/s3/esame venerdi/backoffice.js
@@ -6,7 +6,7 @@ let imagineInput = document.querySelector('#imgUrl')
 let saveButton = document.querySelector('#save-button')
 let deleteButton = document.querySelector('#delete-button')
 let rowReference   = document.querySelector('#events-container')
-
+let fetchUrl = "https://striveschool-api.herokuapp.com/api/product/";
 console.log(nameInput)
 
 function Phone(name = nameInput.value, description = descriptionInput.value, brand = brandInput.value, price = priceInput.value, image = imagineInput.value) {
@@ -17,8 +17,13 @@ function Phone(name = nameInput.value, description = descriptionInput.value, bra
     this.imageUrl = image;
 }
 
-    let getProducts = function () {
-    fetch("https://striveschool-api.herokuapp.com/api/product/")
+let getProducts = function () {
+    
+fetch("https://striveschool-api.herokuapp.com/api/product/", {
+    headers: {
+    "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVlMDVkNzg4Zjc0MDAwMTQyODc0ODgiLCJpYXQiOjE2ODM4ODM0NzksImV4cCI6MTY4NTA5MzA3OX0.k78ZCoP1rC9euVSQ68bNPaqZ6OAMM-MTQpfZc06pTmk"
+    }
+    })
     .then((res) => {
         if (res.ok) {
             return res.json()
@@ -32,7 +37,7 @@ function Phone(name = nameInput.value, description = descriptionInput.value, bra
           let colTemplate = `
           <div class="col-12 col-md-3">
             <div class="card">
-            <img src="${event.imageUrl}" alt="" class="card-img">
+            <img class= "w-100" style="height:220px" src="${event.imageUrl}" alt="">
               <div class="card-body">
                 <h5 class="card-title">${event.name}</h5>
                 <p class="card-text">
@@ -56,6 +61,10 @@ function Phone(name = nameInput.value, description = descriptionInput.value, bra
     })
 }
 
+window.onload = () => {
+    getProducts()
+}
+
 
 
 
@@ -76,47 +85,24 @@ fetch("https://striveschool-api.herokuapp.com/api/product/", {
 
         .then((data) => {
             console.log('EVENTI IN DB', data)
-            data.forEach((event) => {
-              let colTemplate = `
-              <div class="col-12 col-md-3">
-                <div class="card">
-                  <div class="card-body">
-                    <h5 class="card-title">${event.name}</h5>
-                    <p class="card-text">
-                      ${event.description}
-                    </p>
-                    <p>${new Date(event.time).toLocaleDateString('it-IT')} - ${
-                event.price
-              }€</p>
-                    <a href="./backoffice.html?eventId=${
-                      event._id
-                    }" class="btn btn-primary">MODIFICA</a>
-                  </div>
-                </div>
-              </div>
-              `
-              let rowReference = document.getElementById('events-container') // <div class="row"></div>
-              rowReference.innerHTML += colTemplate
-
-            })
           })
-    .catch((err) => {
+        .catch((err) => {
         console.log(err)
     })    
 
 
     saveButton.addEventListener('click', function (e)  {
-        e.preventDefault()
-        fetch("https://striveschool-api.herokuapp.com/api/product/", {
-        method: "POST",
+        e.preventDefault();
+        fetch(eventId ? fetchUrl + eventId : fetchUrl, {
+        method: eventId ? "PUT" : "POST",
+        body: JSON.stringify(new Phone()),
         headers: {
             "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVlMDVkNzg4Zjc0MDAwMTQyODc0ODgiLCJpYXQiOjE2ODM4ODM0NzksImV4cCI6MTY4NTA5MzA3OX0.k78ZCoP1rC9euVSQ68bNPaqZ6OAMM-MTQpfZc06pTmk",
             "Content-Type": "application/json"    
         },
-        body: JSON.stringify(new Phone())
         })
         .then((res) => {
-                console.log(new Phone)
+            location.assign("./homepage.html")
             console.log(res)
         })
         .catch((err) => {
@@ -124,20 +110,31 @@ fetch("https://striveschool-api.herokuapp.com/api/product/", {
         })    
     })
 
+    let adressBarContent = new URLSearchParams(window.location.search)
+    let eventId = adressBarContent.get('eventId')
+    console.log(eventId , "EVENT ID")
 
-    deleteButton.addEventListener("click", function(e) {
-        e.preventDefault();
-        fetch("https://striveschool-api.herokuapp.com/api/product/", {
-            method: "DELETE",
-            headers: {
-                "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVlMDVkNzg4Zjc0MDAwMTQyODc0ODgiLCJpYXQiOjE2ODM4ODM0NzksImV4cCI6MTY4NTA5MzA3OX0.k78ZCoP1rC9euVSQ68bNPaqZ6OAMM-MTQpfZc06pTmk",
-                "Content-Type": "application/json"    
-    }})
+    if (eventId) {
+        deleteButton.addEventListener('click', () => {
+            fetch("https://striveschool-api.herokuapp.com/api/product/" + eventId , {
+        method: "DELETE",
+        headers: {
+            "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVlMDVkNzg4Zjc0MDAwMTQyODc0ODgiLCJpYXQiOjE2ODM4ODM0NzksImV4cCI6MTY4NTA5MzA3OX0.k78ZCoP1rC9euVSQ68bNPaqZ6OAMM-MTQpfZc06pTmk",
+            "Content-Type": "application/json"    
+        }})
     .then((res) =>{
         console.log(res)
+        if(res.ok) {
+            location.assign("./homepage.html")
+        }
     })
     .catch((err) =>{
-        console.log(err)
-    }
-    )
+        throw new Error("problema")
+        
+    })
 })
+    }
+
+
+
+
